# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Glamur Club",
        "description": "Documentación interactiva de la API REST para el Frontend SPA.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "description": "Autentica al usuario y devuelve un token Sanctum y los roles.",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@glamur.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso"
                    },
                    "401": {
                        "description": "Credenciales incorrectas o no autorizado"
                    }
                },
                "servers": [
                    {
                        "url": "/"
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Invalida el token actual del usuario.",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente"
                    },
                    "401": {
                        "description": "Token inválido o ausente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/"
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Catálogo"
                ],
                "summary": "Obtener todos los productos",
                "description": "Devuelve una lista de productos",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida correctamente"
                    }
                }
            }
        },
        "/api/products/featured": {
            "get": {
                "tags": [
                    "Catálogo"
                ],
                "summary": "Obtener productos destacados (Inteligencia)",
                "description": "Devuelve una recomendación inteligente de los 4 productos más Premium (mayor precio) que actualmente tienen stock.",
                "operationId": "getFeaturedProducts",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Catálogo"
                ],
                "summary": "Obtener un producto por ID",
                "description": "Devuelve los detalles de un producto específico.",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token devuelto por el /login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Autenticación"
        },
        {
            "name": "Catálogo",
            "description": "Catálogo"
        }
    ]
}